aosp/android-mainline: categorize and reorder patches
Divide the patches into the following categories, and put them in this
order:

    Android vendor hooks
    kbuild changes
    arm64 changes
    iommu changes
    core kernel changes
    power management changes
    scheduler changes
    uid_sys_stats support
    cpufreq changes
    mm (memory management) changes
    security changes
    binder changes
    ion changes
    driver framework and misc driver changes
    tty changes
    USB changes
    graphics changes
    sound changes
    media changes
    input and HID changes
    UFS driver changes
    MMC driver changes
    block changes
    device-mapper changes
    filesystem changes
    Incremental filesystem
    networking changes
    hardware-wrapped key support
    db845c (DragonBoard 845c) support
    Android kconfigs and build scripts

There's a special "annotation" patch to mark the beginning of each
category.  Note, these maybe should be comments in the 'series' file
instead, but that doesn't work well with importing/exporting the patches
to/from git and using 'git rebase'.  Suggestions appreciated...

The category with the most patches is the last one, which contains all
the gki_defconfig, Kconfig.gki, and build.* changes.  In a few cases,
patches changed both these Android-specific files and upstream kernel
files, so I split those patches up.

Change-Id: I235265bd15cb3bd8c70d9e55ce5bd964144e9011
Signed-off-by: Eric Biggers <ebiggers@google.com>